Backend version 1.0
Finalicé los crud de autor, cliente, libro y prestamo. Todavia quedan validaciones por agregar, y posiblemente agregue mas metodos a los controllers, pero por ahora el backend es funcional.

+++ backend/src/main/java/com/example/apiLibros/Libro/LibroController.java
@@ -51,12 +51,14 @@ public Libro updateLibro(@PathVariable Long id, @RequestBody Libro lib) {
 
 @DeleteMapping("/{id}")
 public String deleteLibro(@PathVariable Long id) {
-    try {
-        libroRepository.findById(id).get();
+    if (libroRepository.existsById(id) )
+    {
         libroRepository.deleteById(id);
         return "Libro eliminado con éxito.";
-    } catch (Exception e) {
-        return "El libro no pudo ser encontrado.";
+    }
+        else
+        {
+            return "El libro no existe.";
         }
     }
 }
